@@ -1,60 +1,70 @@
 {
     "name": "@youwol/fv-button",
-    "keywords": [
-        "html",
-        "reactive-programming",
-        "rxjs",
-        "view",
-        "button"
-    ],
-    "version": "0.0.4",
-    "description": "Button widget using flux-view",
-    "author": "Guillaume Reinisch",
-    "license": "MIT",
-    "homepage": "",
+    "description": "Button's type of widgets using flux-view.",
+    "version": "0.1.0",
+    "author": "greinisch@youwol.com",
+    "homepage": "https://github.com/youwol/fv-button#README.md",
     "main": "dist/@youwol/fv-button.js",
     "types": "src/index.ts",
     "scripts": {
+        "clean": "del-cli dist",
+        "auto-gen": "python template.py",
         "build": "yarn build:dev",
-        "build:dev": "webpack --mode development",
-        "build:prod": "webpack --mode production",
-        "test": "jest -c ./jest.config.js --rootDir .",
-        "test-coverage": "jest -c ./jest.config.js --rootDir . --collect-coverage --collectCoverageFrom=./src/lib/**/*.ts --coverageDirectory=./coverage/$TARGET --runInBand",
+        "pre-build": "yarn clean && yarn auto-gen",
+        "build:dev": "yarn pre-build && webpack --mode development",
+        "build:prod": "yarn pre-build && webpack --mode production",
+        "lint-check": "yarn lint-prettier-check && yarn lint-eslint-check",
+        "lint-prettier-check": "prettier --check ./src",
+        "lint-eslint-check": "eslint ./src",
+        "test": "jest -c ./jest.config.js -i --rootDir .",
+        "test-coverage": "jest -c ./jest.config.js -i --rootDir . --collect-coverage --collectCoverageFrom=./src/lib/**/*.ts ",
+        "doc": "typedoc",
         "publish-npm-public": "yarn build:prod && yarn test-coverage && yarn publish --access public"
     },
-    "repository": {
-        "type": "git",
-        "url": ""
-    },
-    "dependencies": {},
-    "peerDependencies": {
-        "@youwol/flux-view": "^0.0.11",
-        "rxjs": "6.x"
-    },
     "devDependencies": {
-        "@types/node": "14.14.16",
-        "awesome-typescript-loader": "5.2.1",
-        "eslint": "7.16.0",
-        "eslint-loader": "4.0.2",
-        "jest": "26.6.3",
-        "jest-junit": "12.0.0",
-        "rxjs": "6.x",
-        "@youwol/flux-view": "0.x",
-        "rxjs-spy": "7.5.3",
-        "ts-jest": "26.4.4",
-        "typescript": "4.1.3",
-        "webpack": "^4.12.0",
-        "webpack-cli": "4.3.0",
-        "webpack-bundle-analyzer": "4.5.0"
+        "@types/node": "18.0.6",
+        "typescript": "^4.7.4",
+        "ts-lib": "^0.0.5",
+        "ts-node": "^10.9.1",
+        "ts-loader": "^9.3.1",
+        "@types/jest": "27.4.1",
+        "ts-jest": "27.1.3",
+        "jest": "27.5.1",
+        "jest-junit": "13.0.0",
+        "typedoc": "^0.23.8",
+        "webpack": "^5.73.0",
+        "webpack-cli": "4.9.2",
+        "webpack-bundle-analyzer": "4.5.0",
+        "@types/webpack": "^5.28.0",
+        "del-cli": "^5.0.0",
+        "@youwol/prettier-config": "^0.0.1",
+        "@youwol/eslint-config": "^0.0.1",
+        "prettier": "^2.7.1",
+        "prettier-plugin-properties": "^0.2.0",
+        "@typescript-eslint/eslint-plugin": "^5.33.1",
+        "@typescript-eslint/parser": "^5.33.1",
+        "eslint": "8.22.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-eslint-comments": "^3.2.0",
+        "eslint-plugin-import": "^2.26.0",
+        "eslint-plugin-jest": "^26.8.4",
+        "eslint-plugin-tsdoc": "^0.2.16",
+        "eslint-plugin-unused-imports": "^2.0.0"
+    },
+    "prettier": "@youwol/prettier-config",
+    "eslintConfig": {
+        "extends": [
+            "@youwol"
+        ]
+    },
+    "dependencies": {
+        "rxjs": "^6.5.5",
+        "@youwol/flux-view": "^1.0.0"
     },
     "youwol": {
-        "pipeline": {
-            "name": "yw_pipeline_webpack_ts",
-            "version": "0.0.0"
-        },
-        "cdnDependencies":{
-            "@youwol/flux-view": "0.x",
-            "rxjs": "6.x"
+        "cdnDependencies": {
+            "rxjs": "^6.5.5",
+            "@youwol/flux-view": "^1.0.0"
         }
     }
 }
