@@ -1,6 +1,7 @@
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'jsdom',
-  reporters: [ "default", "jest-junit"],
-  modulePathIgnorePatterns : ["/dist","modules-store@youwol"],
-};
+    preset: 'ts-jest',
+    testEnvironment: 'jsdom',
+    testURL: 'http://localhost:2001',
+    reporters: ['default', 'jest-junit'],
+    modulePathIgnorePatterns: ['/dist'],
+}
