@@ -1,6 +1,3 @@
-/*
- * Public API Surface of flux-lib-core
- */
-
- export {Button} from './lib/button.view'
- export {Switch} from './lib/switch.view'
+export { Button } from './lib/button.view'
+export { Switch } from './lib/switch.view'
+export { setup } from './auto-generated'
